LidarTileFootprint modification
Modified LidarTileFootprint so that the bounding box of each input file is identified by default and the convex hull is optional. This significantly improves the efficiency when applied to hundreds or thousands of tiles since 1) the bounding box only requires reading the LAS header, and 2) it does not require computing the convex hull.