Some changes for surface shape metrics in geographic coordinates
Updated the way the z_factor is used to adjust DEM elevations for rasters in geographic coordinates in the surface shape metric tools, e.g. slope, aspect, hillshade, and the curvatures. This allows for the adjustment factor to be calculated per row, rather than using a single raster-wide value based on the middle latitude.