Changes to surface morphology tools
Added Gaussian curvature, minimal curvature and maximal curvature tools. I have also changed the algorithm used to calculate profile, tangental, plan, and mean curvatures as well as slope and aspect. The new algorithm is based on Florinsky's (2016) 5x5 method for projected DEMs, and his 3x3 method described for DEMs in geographic coordinates. This is a big change and we're still in the process of testing...